+++ packages/components/pie-assistive-text/src/index.ts
@@ -3,7 +3,7 @@ import {
 } from 'lit';
 import { PieElement } from '@justeattakeaway/pie-webc-core/src/internals/PieElement';
 import { property } from 'lit/decorators.js';
-import { validPropertyValues, defineCustomElement } from '@justeattakeaway/pie-webc-core';
+import { safeCustomElement, validPropertyValues } from '@justeattakeaway/pie-webc-core';
 import { classMap } from 'lit/directives/class-map.js';
 import '@justeattakeaway/pie-icons-webc/dist/IconAlertCircle.js';
 import '@justeattakeaway/pie-icons-webc/dist/IconCheckCircle.js';
@@ -20,6 +20,7 @@ const componentSelector = 'pie-assistive-text';
  * @tagname pie-assistive-text
  * @slot - Default slot
  */
+@safeCustomElement('pie-assistive-text')
 export class PieAssistiveText extends PieElement implements AssistiveTextProps {
     @property({ type: String })
     @validPropertyValues(componentSelector, variants, defaultProps.variant)
@@ -63,8 +64,6 @@ export class PieAssistiveText extends PieElement implements AssistiveTextProps {
     // Renders a `CSSResult` generated from SCSS by Vite
     static styles = unsafeCSS(styles);
 }
-
-defineCustomElement(componentSelector, PieAssistiveText);
 
 declare global {
     interface HTMLElementTagNameMap {