+++ packages/components/pie-switch/src/index.ts
@@ -8,7 +8,8 @@ import { ifDefined } from 'lit/directives/if-defined.js';
 import 'element-internals-polyfill';
 
 import {
-    RtlMixin, validPropertyValues, defineCustomElement, FormControlMixin, wrapNativeEvent, type PIEInputElement,
+    RtlMixin, validPropertyValues, FormControlMixin, wrapNativeEvent, type PIEInputElement,
+    safeCustomElement,
 } from '@justeattakeaway/pie-webc-core';
 import '@justeattakeaway/pie-icons-webc/dist/IconCheck.js';
 
@@ -24,6 +25,7 @@ const componentSelector = 'pie-switch';
  * @tagname pie-switch
  * @event {CustomEvent} change - when the switch checked state is changed.
  */
+@safeCustomElement('pie-switch')
 export class PieSwitch extends FormControlMixin(RtlMixin(PieElement)) implements SwitchProps, PIEInputElement {
     @property({ type: String })
     public label: SwitchProps['label'];
@@ -225,8 +227,6 @@ export class PieSwitch extends FormControlMixin(RtlMixin(PieElement)) implements
             </label>`;
     }
 }
-
-defineCustomElement(componentSelector, PieSwitch);
 
 declare global {
     interface HTMLElementTagNameMap {