+++ packages/components/pie-breadcrumb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @justeattakeaway/pie-breadcrumb
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`07b8fca`](https://github.com/justeattakeaway/pie/commit/07b8fca26662f794e8e5f6137b1ee7c7f755f6b0)]:
+  - @justeattakeaway/pie-webc-core@0.26.0
+
 ## 0.3.0
 
 ### Minor Changes